indicate accountability
Will need to edit final draft for this to be strictly true, which is the intent; recent remark about LR vs. DL (or, for that matter, RF/similar ML) is worth emphasizing.  If the training data is insufficiently informative, or lacks structure, the marginal benefit of DL relative to the marginal effort involved training DNN/CNN type models is usually difficult to justify.